Added Support To Sync from Cloud Server to Local
Earlier, if we restore cloud to local using mysql dump,
there's a possibility that some rows are duplicated because
synchronizations table only works one-way. When mysql dump from
cloud is restored to local, cloudId becomes localId, hence when
synced, it's no longer consistent.

To solve this, we added onCloud column to tables that are
to be synchronized. This column indicates that the data is
already on the cloud. Hence won't be sent.