Updated VideoService
Previously, VideoService's getAllVideos() returns all videos,
even those that are already replaced with newer ones, because
when there's a new upload for a specific subtopicId, we don't
delete previous entry (for backup & history purposes).

The problem is that video-and-image-sync.js script uses that
function to get the list of videos to be downloaded, and hence all
the past videos are also downloaded, bloating up the size of download.

This changes fixes it by introducing a new SQL view, subtopicVideosView,
which returns only the latest version of each subtopics' videos.